Remove time zone tests
We are just going to rely on AR to do the right thing. It doesn't really, since it only formats the date/time and doesn't actually put it in the right zone, but that is ok.

At GitHub, our custom fork of Rails does make these tests properly and we have integration tests to verify that so we don't need these tests here. We will leave the force_timezone option and related code though because it is necessary for us.