GitHub => Github
Sorry branding, but with GitHub the rails generator was
`git_hub:data:active_record` and it did not work on a brand new rails
app. When I changed it to Github, the generator showed up as
`github:data:active_record` and worked like a charm.

Happy to change this if anyone knows how but I'm too lazy to go farther
since it works for now.