Require the top library
This is me guessing, but I think with the rename (the last rename, and
the really last rename) the console script didn't get updated to
require the "github/ds" instead still requiring "github/kv" (which
also works).

This rewords that, the console still works (basically exactly the same
as before) but it is slightly more correct.