Move time zone enforcement to quoting of dates/times
This makes it work for `interpolate` and `sanitize` when used outside of
`add`. Without this, those will both not force tz.

I don't think this needs to be wrapped around any other connection
quotes as it only relates to dates/times, but let me know if I'm wrong.

I left the enforce timezone in results because those use connection as
well and wasn't 100% confident that I could remove without issue. Likely
it would be fine and I can look into it later.