Don't de-dup values in results array
Problem:

If a user passed duplicate column names to GitHub::DS, for example
`sql.add(SELECT id, NULL, NULL, name from repositories)` the valued
returned by `result` wouldn't line up with the column names. Instead
what would be returned is:

```
sql.results
[[1, nil, "repo name"], [2, nil, "repo name 2"]]
```

instead of the correct value set:

```
sql.results
[[1, nil, nil, "repo name"], [2, nil, nil, "repo name 2"]]
```

This is because we were calling `to_ary` on the
`ActiveRecord::Base#result`. This would loop through the columns and
rows and pair them up into a hash. The underlying problem is hashes
cannot contain keys with the same value, and Ruby will de-dup the keys
before retunring the result.

Solution:

To fix this problem we can instead call `rows` on the returned results
which will give us all the rows. If we want `hash_results` instead we
can accept that a hash de-dups the values since we likely will never
turn a hash back into a SQL statement.

Originally I considered raising if there were duplicate columns passed
in but after some investigation into Rails I found that Rails is able to
take `["title", "title"]` as the columns and then apply those columns
properly to the joined tables, so that wasn't a viable solution.

After considering the options it makes sense for results to return ALL
the results and for the hash to return just the de-duplicated keys. This
is the behavior in Rails for many years so this behavior seems correct.